+++ BLCMM/src/blcmm/gui/tree/rightmouse/InsertAction.java
@@ -44,7 +44,7 @@ import javax.swing.tree.TreePath;
 public class InsertAction extends RightMouseButtonAction {
 
     public InsertAction(CheckBoxTree tree, int hotkey, boolean ctrl) {
-        super(tree, "Insert", hotkey, ctrl, new Requirements(true, true, true, true));
+        super(tree, "Insert", hotkey, ctrl, new Requirements(true, true));
     }
 
     @Override
@@ -62,7 +62,7 @@ public class InsertAction extends RightMouseButtonAction {
             return false;
         }
 
-        return Options.INSTANCE.isInDeveloperMode() && Options.INSTANCE.getStructuralEdits();
+        return true;
 
     }