+++ BLCMM/src/blcmm/gui/tree/rightmouse/PasteAction.java
@@ -48,7 +48,7 @@ public class PasteAction extends RightMouseButtonAction {
     private DefaultMutableTreeNode parentNode;
 
     public PasteAction(CheckBoxTree tree, int hotkey, boolean ctrl) {
-        super(tree, "Paste", hotkey, ctrl, new Requirements(true, false, true, false));
+        super(tree, "Paste", hotkey, ctrl, new Requirements(true, false));
     }
 
     @Override